% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingDist.R
\name{samplingDist}
\alias{samplingDist}
\title{Sampling Distribution}
\usage{
samplingDist(samples = 50, sample_size = 100, mean = 0, sd = 1)
}
\arguments{
\item{samples}{Specify how many samples to run}

\item{sample_size}{the number of observations in each sample}

\item{mean}{population mean}

\item{sd}{population standard deviation}
}
